# Try to optimize `LazyLoadTimelines`