--- a/state/event_table.go
+++ b/state/event_table.go
@@ -345,6 +345,7 @@ func (t *EventTable) SelectLatestEventsBetween(txn *sqlx.Tx, roomID string, lowe
 	return events, err
 }
 
+// / SELECT prev_batch FROM syncv3_events WHERE prev_batch IS NOT NULL AND room_id=$1 AND event_nid >= $2 LIMIT 1
 func (t *EventTable) SelectLatestEventsBetweenV2(txn *sqlx.Tx, roomIDs []string, startNIDs, endNIDs []int64, limit int) ([]Event, error) {
 	var events []Event
 	// do not pull in events which were in the v2 state block
@@ -355,8 +356,8 @@ WITH input AS (
            unnest($3::bigint[]) end
 )
 SELECT room_id, event_nid, event, prev_batch FROM input,
-lateral ( SELECT  room_id, event_nid, event FROM syncv3_events WHERE event_nid > input.start AND event_nid <= input.end AND room_id = input.roomID AND is_state=FALSE ORDER BY event_nid DESC LIMIT $4 ) AS events,
-lateral (SELECT prev_batch FROM syncv3_events WHERE prev_batch IS NOT NULL AND room_id=input.roomID AND event_nid <= input.end ORDER BY event_NID DESC LIMIT 1) AS prevBatch
+lateral (SELECT room_id, event_nid, event FROM syncv3_events WHERE event_nid > input.start AND event_nid <= input.end AND room_id = input.roomID AND is_state=FALSE ORDER BY event_nid DESC LIMIT $4 ) AS events,
+lateral (SELECT prev_batch FROM syncv3_events WHERE prev_batch IS NOT NULL AND room_id=input.roomID AND event_nid >= events.event_nid LIMIT 1) AS prevBatch
 `,
 		pq.StringArray(roomIDs), pq.Int64Array(startNIDs), pq.Int64Array(endNIDs), limit,
 	)