--- a/sync3/caches/user.go
+++ b/sync3/caches/user.go
@@ -308,11 +308,13 @@ func (c *UserCache) OnRegistered(ctx context.Context) error {
 
 // Load timelines from the database. Uses cached UserRoomData for metadata purposes only.
 func (c *UserCache) LazyLoadTimelines(ctx context.Context, loadPos int64, roomIDs []string, maxTimelineEvents int) map[string]UserRoomData {
+	ctx, span := internal.StartSpan(ctx, "LazyLoadTimelines")
+	defer span.End()
 	if c.LazyRoomDataOverride != nil {
 		return c.LazyRoomDataOverride(loadPos, roomIDs, maxTimelineEvents)
 	}
 	result := make(map[string]UserRoomData)
-	roomIDToLatestEvents, err := c.store.LatestEventsInRooms(c.UserID, roomIDs, loadPos, maxTimelineEvents)
+	roomIDToLatestEvents, err := c.store.LatestEventsInRoomsV2(c.UserID, roomIDs, loadPos, maxTimelineEvents)
 	if err != nil {
 		logger.Err(err).Strs("rooms", roomIDs).Msg("failed to get LatestEventsInRooms")
 		internal.GetSentryHubFromContextOrDefault(ctx).CaptureException(err)
@@ -408,6 +410,8 @@ func (c *UserCache) Invites() map[string]UserRoomData {
 // which would cause the transaction ID to be missing from the event. Instead, we always look for txn
 // IDs in the v2 poller, and then set them appropriately at request time.
 func (c *UserCache) AnnotateWithTransactionIDs(ctx context.Context, userID string, deviceID string, roomIDToEvents map[string][]json.RawMessage) map[string][]json.RawMessage {
+	ctx, span := internal.StartSpan(ctx, "AnnotateWithTransactionIDs")
+	defer span.End()
 	var eventIDs []string
 	eventIDToEvent := make(map[string]struct {
 		roomID string